pyproject.toml
[project]
name = "second-brain-ocr"
version = "0.1.0"
description = "Automated OCR and vector search indexing for your second brain notes"
requires-python = ">=3.13"
dependencies = [
    "azure-ai-documentintelligence>=1.0.2",
    "azure-ai-formrecognizer>=3.3.3",
    "azure-identity>=1.25.1",
    "azure-search-documents>=11.6.0",
    "openai>=2.3.0",
    "python-dotenv>=1.1.1",
    "watchdog>=6.0.0",
]

[project.scripts]
second-brain-ocr = "second_brain_ocr.main:main"

[tool.setuptools.package-data]
second_brain_ocr = ["py.typed"]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.0",
]
